ordering_agent/mcp_server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
authlib==1.6.1
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
cyclopts==3.22.2
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.17.0
docutils==0.21.2
email-validator==2.2.0
exceptiongroup==1.3.0
fastmcp==2.10.6
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mcp==1.12.1
mdurl==0.1.2
numpy==2.3.1
openai==1.97.1
openapi-pydantic==0.5.1
portalocker==3.2.0
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyperclip==1.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
qdrant-client==1.15.0
referencing==0.36.2
rich==14.0.0
rich-rst==1.3.1
rpds-py==0.26.0
sniffio==1.3.1
sse-starlette==2.4.1
starlette==0.47.2
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
